% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{field2}
\alias{field2}
\title{Field 2 data: Climate and rice blast observations from Kilinochchi, Sri Lanka}
\format{
A data frame with 2880 rows and 9 variables:
\describe{
\item{time}{Time Stamps}
\item{d2m}{2m dewpoint temperature. Temperature where air moisture starts to condense at 2m height.}
\item{t2m}{Air temperature 2 meters above ground.}
\item{stl1}{Soil temperature level 1. Temperature of the top soil layer.}
\item{u10}{10m u-component of wind: East-west wind speed at 10m.(positive means wind blowing east).}
\item{v10}{10m v-component of wind: North-south wind speed at 10m.(positive means wind blowing north).}
\item{tp}{Total precipitation: The amount of rain, snow, or other moisture that falls.}
\item{type}{Whether individual data points are disease points or not. "typical", "disease"}
\item{RH}{Relative Humidity: How much moisture is in the air compared to its max capacity.
Calculated using d2m and t2m. Calculated using RH <- 100 * (exp((17.625 * d2m) / (243.04 + d2m))) / (exp((17.625 * t2m) / (243.04 + t2m)))}}
}
\source{
Historical climate data obtained from the Copernicus EU Earth Observation Programme <https://www.copernicus.eu/en>, and rice blast disease data compiled with expert guidance from the Department of Agriculture Sri Lanka.
}
\usage{
field2
}
\description{
This dataset contains historical climate and rice blast disease observations
for Field 2 located in Kilinochchi, Sri Lanka (latitude 9.313077, longitude 80.294791).
At this site, planting occurred on 2020-10-15, and the disease was observed on 2021-02-12.
The data were collected from the Copernicus EU Earth Observation Program (https://www.copernicus.eu/en),
which integrates satellite and ground-based information to provide global climate coverage,
and supplemented with rice blast disease records obtained from the Department of Agriculture,
Sri Lanka. Field visits were conducted during the 2020/2021 “Maha” season when blast
outbreaks were reported. Farmers observed that under favorable environmental conditions,
complete field infection occurred within 24–48 hours after the appearance of initial symptoms
}
\keyword{datasets}
